refactor: Extract util/fs from util/system
