+++ src/addrdb.cpp
@@ -19,6 +19,7 @@
 #include <streams.h>
 #include <tinyformat.h>
 #include <univalue.h>
+#include <util/fs.h>
 #include <util/settings.h>
 #include <util/system.h>
 #include <util/translation.h>
