+++ src/dbwrapper.cpp
@@ -8,6 +8,7 @@
 #include <logging.h>
 #include <random.h>
 #include <tinyformat.h>
+#include <util/fs.h>
 #include <util/strencodings.h>
 #include <util/system.h>
