+++ src/flatfile.cpp
@@ -8,6 +8,7 @@
 #include <flatfile.h>
 #include <logging.h>
 #include <tinyformat.h>
+#include <util/fs.h>
 #include <util/system.h>
 
 FlatFileSeq::FlatFileSeq(fs::path dir, const char* prefix, size_t chunk_size) :
