+++ src/index/blockfilterindex.cpp
@@ -9,6 +9,7 @@
 #include <index/blockfilterindex.h>
 #include <logging.h>
 #include <node/blockstorage.h>
+#include <util/fs.h>
 #include <util/system.h>
 #include <validation.h>
