+++ src/init.cpp
@@ -69,6 +69,7 @@
 #include <txmempool.h>
 #include <util/asmap.h>
 #include <util/check.h>
+#include <util/fs.h>
 #include <util/moneystr.h>
 #include <util/shell.h>
 #include <util/strencodings.h>
