+++ src/qt/guiutil.cpp
@@ -21,6 +21,7 @@
 #include <script/script.h>
 #include <script/standard.h>
 #include <util/exception.h>
+#include <util/fs.h>
 #include <util/system.h>
 #include <util/time.h>
