+++ src/qt/intro.cpp
@@ -16,6 +16,7 @@
 #include <qt/optionsmodel.h>
 
 #include <interfaces/node.h>
+#include <util/fs.h>
 #include <util/system.h>
 #include <validation.h>
