+++ src/qt/walletframe.cpp
@@ -12,6 +12,7 @@
 #include <qt/psbtoperationsdialog.h>
 #include <qt/walletmodel.h>
 #include <qt/walletview.h>
+#include <util/fs.h>
 #include <util/system.h>
 
 #include <cassert>
