+++ src/rpc/request.cpp
@@ -9,6 +9,7 @@
 
 #include <random.h>
 #include <rpc/protocol.h>
+#include <util/fs.h>
 #include <util/system.h>
 #include <util/strencodings.h>
