+++ src/test/fs_tests.cpp
@@ -4,7 +4,7 @@
 //
 #include <fs.h>
 #include <test/util/setup_common.h>
-#include <util/system.h>
+#include <util/fs.h>
 #include <util/getuniquepath.h>
 
 #include <boost/test/unit_test.hpp>
