+++ src/validation.cpp
@@ -46,6 +46,7 @@
 #include <uint256.h>
 #include <undo.h>
 #include <util/check.h> // For NDEBUG compile time check
+#include <util/fs.h>
 #include <util/hasher.h>
 #include <util/moneystr.h>
 #include <util/rbf.h>
