+++ src/wallet/bdb.cpp
@@ -8,6 +8,7 @@
 #include <wallet/bdb.h>
 #include <wallet/db.h>
 
+#include <util/fs.h>
 #include <util/strencodings.h>
 #include <util/translation.h>
