+++ src/wallet/sqlite.cpp
@@ -9,6 +9,7 @@
 #include <logging.h>
 #include <sync.h>
 #include <util/strencodings.h>
+#include <util/fs.h>
 #include <util/system.h>
 #include <util/translation.h>
 #include <wallet/db.h>
